refactor(macos): extract Snapshot3D enter/exit JS into bundled resources
Moved the 800-line 3D-inspector script blobs out of Swift string literals
and into `Resources/Snapshot3D/*.js`. The enter script is split along the
phase comments that the original author already used (setup/collect/
apply/input), each under 500 lines. Concatenation is byte-identical to
the previous single string, so the running script and its shared IIFE
closure scope are unchanged.

before: Snapshot3DBridge.swift 802 lines
after:
  Snapshot3DBridge.swift                       70
  Resources/Snapshot3D/enter-setup.js         108
  Resources/Snapshot3D/enter-collect.js       141
  Resources/Snapshot3D/enter-apply.js         111
  Resources/Snapshot3D/enter-input.js         342
  Resources/Snapshot3D/exit.js                 65
